+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/impex/ImpexUnknownTypeNameInspection.kt
@@ -1,5 +1,5 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
@@ -44,10 +44,7 @@ class ImpexUnknownTypeNameInspection : LocalInspectionTool() {
         }
 
         override fun visitAnyAttributeValue(element: ImpexAnyAttributeValue) {
-            element.anyAttributeName
-                ?.text
-                ?.takeIf { it == TypeModifier.DISABLE_UNIQUE_ATTRIBUTES_VALIDATOR_FOR_TYPES.modifierName }
-                ?: return
+            if (TypeModifier.getModifier(element) != TypeModifier.DISABLE_UNIQUE_ATTRIBUTES_VALIDATOR_FOR_TYPES) return
 
             validateReference(element)
         }