+++ src/com/intellij/idea/plugin/hybris/impex/completion/provider/ImpexHeaderTypeModifierValueCompletionProvider.kt
@@ -1,7 +1,7 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -43,7 +43,7 @@ class ImpexHeaderTypeModifierValueCompletionProvider : CompletionProvider<Comple
         val psiElementUnderCaret = parameters.position
         val impexAttribute = PsiTreeUtil.getParentOfType(psiElementUnderCaret, ImpexAttribute::class.java) ?: return
         val modifierName = impexAttribute.anyAttributeName.text
-        val impexModifier = TypeModifier.getByModifierName(modifierName)
+        val impexModifier = TypeModifier.getModifier(modifierName)
 
         if (impexModifier != null) {
             impexModifier.getLookupElements(project)