+++ src/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexAttributeValueMixin.kt
@@ -1,5 +1,5 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
@@ -50,6 +50,8 @@ abstract class ImpexAttributeValueMixin(astNode: ASTNode) : ASTWrapperPsiElement
         ?.let { arrayOf(it) }
         ?: computeReference()
 
+    override fun getReference() = references.firstOrNull()
+
     private fun computeReference() = if (getParentOfType<ImpexFullHeaderTypeImpl>(false) != null) {
         computeTypeModifierReference()
     } else {
@@ -63,7 +65,7 @@ abstract class ImpexAttributeValueMixin(astNode: ASTNode) : ASTWrapperPsiElement
         val modifierName = (parent as? ImpexAttribute)
             ?.anyAttributeName
             ?.text
-            ?.let { TypeModifier.getByModifierName(it) }
+            ?.let { TypeModifier.getModifier(it) }
             ?: return null
 
         return when (modifierName) {