+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 ### `ImpEx` inspection rules
 - Inspect type set for `disable.UniqueAttributesValidator.for.types` type modifier [#896](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/896)
 - Use inverted comma for attribute modifiers [#899](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/899)
+- Inspect a value set for `processor` modifier [#900](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/900)
 
 ### `FlexibleSearch` enhancements
 - Improved `words case` and `table alias` actions performance [#874](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/874)